+++ test/UserRepository/HtpasswdTest.php
@@ -1,20 +1,21 @@
 <?php
-/**
- * @see       https://github.com/zendframework/zend-expressive-authentication for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
- */
-namespace ZendTest\Expressive\Authentication\UserRepository;
 
+/**
+ * @see       https://github.com/mezzio/mezzio-authentication for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authentication/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authentication/blob/master/LICENSE.md New BSD License
+ */
+namespace MezzioTest\Authentication\UserRepository;
+
+use Mezzio\Authentication\UserInterface;
+use Mezzio\Authentication\UserRepository\Htpasswd;
+use Mezzio\Authentication\UserRepositoryInterface;
 use PHPUnit\Framework\TestCase;
-use Zend\Expressive\Authentication\UserInterface;
-use Zend\Expressive\Authentication\UserRepositoryInterface;
-use Zend\Expressive\Authentication\UserRepository\Htpasswd;
 
 class HtpasswdTest extends TestCase
 {
     /**
-     * @expectedException Zend\Expressive\Authentication\Exception\InvalidConfigException
+     * @expectedException Mezzio\Authentication\Exception\InvalidConfigException
      */
     public function testConstructorWithNoFile()
     {
@@ -49,7 +50,7 @@ class HtpasswdTest extends TestCase
     }
 
     /**
-     * @expectedException Zend\Expressive\Authentication\Exception\RuntimeException
+     * @expectedException Mezzio\Authentication\Exception\RuntimeException
      */
     public function testAuthenticateWithInsecureHash()
     {